requirements.txt
alembic==1.8.1
amqp==5.1.1
anyio==3.6.1
arrow==1.2.3
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.3
colorama==0.4.5
deprecated==1.2.13
fastapi-event==0.1.3
fastapi==0.81.0
greenlet==1.1.3
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
kombu==5.2.4
mako==1.2.2
markupsafe==2.1.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.4
py==1.11.0
pydantic==1.10.1
pyjwt==2.4.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pythondi==1.2.4
pytz==2022.2.1
pyyaml==6.0
redis==4.3.4
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
sqlalchemy==1.4.40
sqlalchemy[asyncio]==1.4.40
starlette==0.19.1
tomli==2.0.1
typing-extensions==4.3.0
ujson==5.4.0
uvicorn==0.19.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
